Macos support. (#3)
* Macos patching support.

* Unify all patching paths.

* Rename Patcher to AssemblyProvider.

* Switch to MachoSharp for macOS.

* Add an optional output path argument.

Very useful for testing and not continually needing to replace my binary with a backup.

* Add ad-hoc signing for macos, using codesign or quill.

Implements ad-hoc signing (non-cryptographic) of the macos executable after patching. To implement this we either require `codesign` from the Xcode cli tools, or Quill (available on all platforms).

* Paths may have spaces.

* Fix codesign detection.

Codesign doesn't have a --help command or something we can easily test for exit code 0, so we exec and check for the start of the help message.

* Some minor cleanup.

* Address review comments.

* Move signing error to exception message.